+++ packages/http/src/types/webhook.d.ts
@@ -1,4 +1,4 @@
-export type Platform = "bili-recorder" | "blrec" | "ddtv" | "custom";
+export type Platform = "bili-recorder" | "blrec" | "ddtv" | "onelivrec" | "custom";
 export type PickPartial<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>> & Partial<Pick<T, K>>;
 export type UploadStatus = "pending" | "uploading" | "uploaded" | "error";
 export type OpenEvent = "FileOpening";
@@ -26,7 +26,7 @@ export interface Part {
 export interface Options {
   event: OpenEvent | CloseEvent | ErrorEvent;
   filePath: string;
-  roomId: number;
+  roomId: string;
   time: string;
   username: string;
   title: string;
@@ -41,7 +41,7 @@ export interface CustomEvent {
   /** 视频文件的绝对路径 */
   filePath: string;
   /** 房间号，用于断播续传需要 */
-  roomId: number;
+  roomId: string;
   /** 用于标题格式化的时间，示例："2021-05-14T17:52:54.946" */
   time: string;
   /** 标题，用于格式化视频标题 */